Pass configuration to pager in environment
Instead of passing arguments to less on the command line, pass them in
flags in the environment using the LESS environment variable. This
allows the user to override the arguments by setting the LESS
environment variable.

This is the strategy that git uses. It also ensures that if the pager
command is provided by the user via the PAGER environment variable, we
can still pass arguments to it.