Always flush output and handle broken pipe
Previously we only flushed the output stream (by closing it) and
suppressed BrokenPipe exceptions if we were writing to a pager. However,
in practice, all programs will want consistent behaviour regardless of
whether the output is going to the pager or being piped to something
else (which could even be a pager invoked manually) that has exited.

Always flush the output stream when exiting the context manager, and
ignore any broken pipe errors. If it is a pipe to a pager subprocess
that we opened, or if the other end of the pipe is already closed, then
also close the stream. If stdout remains open with unflushed data at
exit, then Python itself tries to flush the stream and prints an ugly
error message when that fails due to the broken pipe, with no way for
the program to control that.

If the output stream is to a broken pipe, suppress any BrokenPipe
exceptions that occurred while the context manager was active,
regardless of whether we invoked a pager.

Ideally we would only suppress the exception when flushing the output
stream reveals a broken pipe, to hopefully avoid suppressing any other
BrokenPipeErrors from unrelated streams whose behaviour should be
unaffected by the presence of the AutoPager. However, there is no
guarantee that the buffer contains anything and therefore the flush()
may not fail even if a previous write failed due to a broken pipe error.