Allow user to pass arguments in $PAGER envvar
If the user wishes to override the settings configured in the
environment, they need to be able to specify a custom PAGER environment
variable containing arguments as well as the name of the pager. Popen
does not support this being passed as a string unless shell=True, but we
generally don't want to be enabling shell interpretation. Therefore
split the arguments using the shlex module.