Don't reset terminal on pager exit by default
Previously, we always allowed less to reset the terminal when exiting
(as it does by default when being run manually), so that the output of
the program did not remain on screen. This is desirable in some
applications, but less so in others. Git does not clear the screen and
instead leaves the last shown output displayed. Change the default to do
the same here.

Since clearing the screen is sometimes desirable, add a configuration
option to allow the developer to choose this.

If we quit the pager immediately due to having less than one screen's
worth of data to display, then the output will obviously always be left
on the screen. This is inconsistent with a request to reset the terminal
on completion, so do not pass this option when the terminal is to be
reset.