Only disable --quit-if-one-screen when line buffering explicitly enabled
By default, Python enables line buffering on stdout whenever stdout is
connected to a console - which is exactly when we will start a pager. As
a consequence, the --quit-if-one-screen option was never enabled unless
the AutoPager was explicitly created with line_buffering=False.

This isn't a great default, given that the programmer knows that the
output will be going to a pager, so only disable --quit-if-one-screen if
line buffering is explicitly turned on.