Use custom SIGINT handler for async
Using the regular handler for SIGINT (which raises KeyboardInterrupt)
with async code doesn't work, because there's no controlling where it is
raised and large parts of the event loop that is unprotected. This means
that it can cause the program to exit even if all of the currently
running tasks have code to catch it. Furthermore, if it does then the
__aexit__() function will be run from a destructor on cleanup after even
the IgnoreHandler has been deleted in CPython, which makes it impossible
to set the handler to SIG_IGN (a cryptic TypeError is raised).

Instead, set a custom handler for the duration of the async context
manager. Use this to cancel the task when SIGINT is received.

Minor downsides of this:
- The signal handler is global, so other threads are also affected.
- The task is now interrupted by CancelledError instead of
  KeyboardInterrupt as expected, and this cannot be changed.
- The whole task in which the context manager is running gets cancelled.