Move subprocess handling to a helper module
A lot of the complexity of the async implementation is related to the
deficiencies of the asyncio.create_subprocess_exec(), such as returning
only binary writers, as compared with the regular subprocess module.

Separate out this handling into a separate AsyncPopen helper class with
an interface similar to what we expect in the synchronous case.