Set SIGPIPE exit code if flush fails
If the output consumer closes before we are able to write all output to
it, then it is irrelevant whether the code inside the context manager
encountered a BrokenPipeError or if it only occurs when trying to flush
the buffer filled by that code. Therefore, set the SIGPIPE return code
if flushing the buffer fails.