Protect all of pager cleanup from SIGINT
It is not enough to catch and handle KeyboardInterrupt while waiting for
the user to exit the pager. A KeyboardInterrupt can occur between any
two Python opcodes. Furthermore, it may (indeed, in some cases it
appears quite likely to) occur while flushing the write buffer and
closing the output stream.

The only way to fully protect against this is to ignore the signal
altogether for the duration of the cleanup.