Add UserSpecifiedPager pager command option
Make it easy for consumers to use custom environment variables instead
(or as well as) the default $PAGER to allow users to specify the pager.
This is a common pattern to allow users to customise the output of
different programs separately, e.g. git uses $GIT_PAGER in preference to
$PAGER if both exist.