Don't read env by default if run with suid bit
If we are running from an executable with the suid bit set, the euid
will be different to the actual uid. Don't allow the pager command to be
read from an environment variable by default in this case. An exception
is if the actual uid is 0 (i.e. root), since this likely denotes a case
where the root user has dropped priveleges rather than the other way
around.