Allow range of types for pager_command parameter
It is too easy to make the mistake of passing e.g. a subclass of
PagerCommand instead of an instance, and while static type checking
would catch the problem there was no runtime checking (which a majority
of consumers are likely to rely on).

Eliminate the problem by accepting either the instance or the class (or
other function that generates an instance. Also allow a string to
generate a CustomPager or a list of strings to pass to
UserSpecifiedPager. Check the type in the constructor for AutoPager so
that any errors will be immediately apparent.