Allow building for EOL Fedora versions
Prior to Fedora 31, the pyproject RPM macros package was not present in
the buildroot, so we cannot build with pyproject.toml (adding it as a
build dependency doesn't help, as it is needed to calculate the build
dependencies). Force them to build the old way, as for CentOS.

In Fedora 31 and 32, the pyproject RPM macros are in the buildroot, but
the latest available version of setuptools doesn't meet the minimum
requirement in pyproject.toml (43.0 - which is set for generating the right
tarball to upload to PyPI, and much higher than what is actually needed
for packaging). Rather than patch the pyproject.toml file, just fall
back to building these the old way as well. The included setup.py file
requires only setuptools 30.3, which is available since Fedora 26.