integration tests: Improve error reporting
Nothing should fail in the first child process (the one created by
pty.fork()), but if it does previously the test would just hang. Make it
easier to debug by returning the stack trace, and open the far end of
all FIFOs so that the parent process doesn't block on them.