integration tests: Ignore sys.stdout when getting tty
After pty.fork(), in the child process it's always the file descriptor 1
that is the controlling tty. This is true even if the test runner
substitutes its own file descriptor for stdout. Therefore, make
ourselves more independent of the test runner by ignoring sys.stdout and
always using the original file descriptor.