tox: Fix dependencies for py36 environment
Ensure the pip version is at least 19.0, because this is required to use
pyproject.toml instead of a setup.py file to install the package to be
tested.

Also explicitly install testtools, because versions of pip prior to 19.0
don't appear to support extras, and thus the fixtures[streams]
dependency doesn't work. pip isn't updated until after all of the
dependencies are installed, and there isn't currently a (working) way to
update it earlier.