Fix integration tests with less >= v613
The behaviour of less when interrupting an input stream with less than a
screenful of data changed subtly in v610 with
https://github.com/gwsw/less/commit/953257f, and later again in v613
with https://github.com/gwsw/less/commit/9d23e5c. This caused the
InvokePagerTest::test_short_streaming_output test to fail.

Where previously less printed the ':' prompt before redrawing the screen
in this scenario, it no longer does. The PagerControl code recognised
the ':' as indicating the end of the output, and prevented the
subsequent redraw being recounted as new ouput. This change ensures that
the redraw will always be counted as the end of the current output.

This means that the integration tests can handle the output given by
either of the releases v608 or v633.

Fixes #5