Fix tty line buffering detection on Illumos
On Illumos (and presumably Solaris), fseek() doesn't return an error
when called on a TTY, contrary to the Solaris man page for e.g.
fflush(), which states that TTYs are not seekable. Consequently, in
Python the seekable() method returns True. This is not the case on
Linux, where fseek() on a TTY results in an error.

To ensure line buffering is correctly enabled on TTY inputs (and prevent
unit test failures), check that the input stream is not a TTY if it
claims to be seekable.

Fixes #4